dairyproject/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
idna==3.4
importlib-metadata==4.12.0
Markdown==3.4.1
mysql-connector==2.2.9
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
wincertstore==0.2
zipp==3.8.1
requests_html==0.10.0
django-redis==5.2.0